[android] filesystem: move implementation to platform folder.
