+++ xbmc/platform/android/filesystem/APKDirectory.cpp
