+++ xbmc/platform/android/filesystem/APKDirectory.h
@@ -19,36 +19,18 @@
  *
  */
 
-#include "IFile.h"
-#include "URL.h"
-
-struct zip;
-struct zip_file;
+#include "filesystem/IFileDirectory.h"
 
 namespace XFILE
 {
-  class CAPKFile : public IFile
+  class CAPKDirectory : public IFileDirectory
   {
-  public:
-    CAPKFile();
-    virtual ~CAPKFile();
-    virtual bool Open(const CURL& url);
-    virtual void Close();
+    public:
+    CAPKDirectory() {};
+    virtual ~CAPKDirectory() {};
+    virtual bool GetDirectory(const CURL& url, CFileItemList &items);
+    virtual bool ContainsFiles(const CURL& url);
+    virtual DIR_CACHE_TYPE GetCacheType(const CURL& url) const;
     virtual bool Exists(const CURL& url);
-
-    virtual int64_t Seek(int64_t iFilePosition, int iWhence = SEEK_SET);
-    virtual ssize_t Read(void* lpBuf, size_t uiBufSize);
-    virtual int Stat(struct __stat64* buffer);
-    virtual int Stat(const CURL& url, struct __stat64* buffer);
-    virtual int64_t GetLength();
-    virtual int64_t GetPosition();
-    virtual int GetChunkSize();
-  protected:
-    CURL              m_url;
-    int               m_zip_index;
-    int64_t           m_file_pos;
-    int64_t           m_file_size;
-    struct zip_file   *m_zip_file;
-    struct zip        *m_zip_archive;
   };
 }
