+++ xbmc/platform/android/filesystem/APKFile.cpp
