+++ xbmc/platform/android/filesystem/AndroidAppFile.cpp
