+++ xbmc/platform/android/filesystem/CMakeLists.txt
@@ -1,0 +1,10 @@
+set(SOURCES APKDirectory.cpp
+            APKFile.cpp
+            AndroidAppDirectory.cpp
+            AndroidAppFile.cpp)
+set(HEADERS APKDirectory.h
+            APKFile.h
+            AndroidAppDirectory.h
+            AndroidAppFile.h)
+
+core_add_library(platform_android_filesystem)
